moved the test from AnalysisRunner to VerificaitionSuite, also fixed a test in Constraint Json string
